+++ package.json
@@ -24,6 +24,7 @@
     "react-router": "^6.18.0",
     "react-router-dom": "^6.18.0",
     "react-scripts": "5.0.1",
+    "react-scroll": "^1.9.0",
     "react-slidy": "^4.3.3",
     "react-spinners": "^0.13.8",
     "typescript": "^4.9.5",
@@ -61,8 +62,9 @@
   },
   "devDependencies": {
     "@mate-academy/eslint-config-react-typescript": "^1.0.12",
-    "@types/react-modal": "^3.16.2",
     "@types/axios": "^0.14.0",
+    "@types/react-modal": "^3.16.2",
+    "@types/react-scroll": "^1.8.10",
     "@typescript-eslint/eslint-plugin": "^5.9.1",
     "@typescript-eslint/parser": "^5.9.1",
     "eslint": "^8.52.0",
