--- src/components/footer.css
@@ -1,77 +1,0 @@
-.footer-content {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: white;
-  margin: 32px;
-}
-
-.footer-logo-cont {
-  width: 89px;
-  height: 32px;
-  object-fit: contain;
-}
-
-.footer-logo {
-  width: 89px;
-  height: 32px;
-  object-fit: contain;
-}
-
-.footer-buttons {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.footer-buttons button,
-.footer-back-to-top button {
-  font-size: 12px;
-  text-decoration: none;
-  padding: 5px 10px;
-  border: none;
-  background-color: white;
-  color: #89939a;
-  cursor: pointer;
-}
-
-.footer-arrow {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 16px;
-  height: 16px;
-}
-.footer-arrow-img {
-  align-items: center;
-  justify-content: center;
-  width: 8px;
-  height: 8px;
-}
-
-.footer-back-to-top {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 32px;
-  height: 32px;
-  border-radius: 48px;
-  border: 1px solid #b4bdc3;
-}
-
-@media screen and (max-width: 639px) {
-  .footer-content {
-    flex-direction: column;
-    align-items: flex-start;
-    background-color: white;
-  }
-
-  .footer-buttons {
-    flex-direction: column;
-  }
-
-  .footer-back-to-top {
-    line-height: 15px;
-    margin-left: 40%;
-  }
-}
